apps/bet-router-rofl/pyproject.toml
[tool.poetry]
name = "bet-router-rofl"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "bet_router_rofl"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
web3 = "^7.12.0"
requests = "^2.32.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
